Fix attribute lookup regression on H5 files
If an attribute in an HDF5v3 data set is not found in the attrs
section of the TelescopeState group, one should check the sensors as
well. When #196 introduced telstate-level decoding, UnpicklingErrors
was replaced by DecodeErrors but unfortunately the KeyError in this
except clause got lost in the process.

This addresses JIRA ticket SR-1503.