Fix applycal crash when doing no selection
The _set_keep method is part of the select() machinery and is supposed
to keep the sensor cache and underlying concatenated datasets in sync
whenever the selection masks change. It is somewhat abused in visdatav4
to also construct new lazy indexers for vis/weights/flags, which
creates and stores new dask graphs.

The problem is that applycal transforms are also added to these dask
graphs, and they need valid corr_products to find the relevant
correction sensors. But corr_products start empty and is only
initialised in the very first select() call, but *after* _set_keep was
called.

The better solution is to rework the whole _set_keep mess, but for now
just postpone _set_keep in the select() call until after corr_products
have been set. It could have been the very last step in select(),
except that scan_indices etc need the sensor cache to be updated first.