Flip the sign of delay correction in applycal
The cal pipeline returns "delays" as a raw phase slope, i.e. a positive
delay implies a positive phase slope. This means that the delays can be
directly applied as delay adjustments in the delay tracker (e.g. in the
calibrate_delays.py observation script), since the delay tracker
implicitly treats a positive delay as a negative phase slope (the usual
definition).

However, when used in applycal to correct visibilities, the delays will
still need a minus sign to become a correction. This was missed due to
the previous reasoning. Also fix the unit test in lockstep.

This addresses JIRA ticket SR-1575.