Change padding strategy for upgrade_chunk_info
The previous strategy was to take the number of dumps in the original
info as golden, and extend or truncate the improved info to fit. That
won't work so well with multi-dump chunks.

Instead, set the number of dumps to be the largest amongst all arrays
(after upgrades), and extend any array that is shorter to ensure a
match. That means that some files may now appear to have extra dumps for
which all the original data is missing, which will then be flagged.

Another side effect is that when not specifying a chunk-store OR
timestamps, the upgrade process is still done just to determine how many
dumps there are.

This should theoretically work with multi-dump chunks, but that has yet
to be tested.