Add max_chunk_sizes option to generate_chunks
This will help with the spectral vis writer, allowing the number of
channels to be limited.

The algorithm was changed to determine an (integral) chunk shape, and then
divide everything into chunks of that shape (with smaller left-overs at
the end).

The behaviour changes slightly: when not using power_of_two, the chunks
are now all the same size except the last which is smaller (like with
power_of_two), rather than having chunks of size X+1 and X. This
simplifies the code, and I don't think !power_of_two is being used much
if at all any more.