Rework reference antenna to fix LST and others
This is a slight rework of the reference antenna machinery to fix
issues with DataSet.lst and a broken ConcatenatedDataSet.

Instead of leaving the `ref_ant` attribute empty in the default case,
explicitly set it to "array". This creates a fake antenna called "array"
with its own katpoint.Antenna object, target sensor and activity sensor.
The rest of the katdal sensor machinery in the base class can then
access the reference antenna properties (such as local sidereal time, or
LST) via the usual mechanisms. It also fixes the bug reported in
SR-1587. In the process, also set up target and activity sensor aliases
for the rest of the antennas to have a consistent view on these.

This addresses JIRA ticket SR-1587.