Fix bug in extraction of array reference
The code pulling the array reference from the first antenna object
did not actually work... When the delay model is zeroed, the antenna
description string changes to report the actual position instead of
the reference position, which will be the cached position of the first
antenna. A more brute force approach chops off the description string
after the first five fields (and changes the name to 'array').