Optimise obtaining corrections from sensor cache
Querying a CategoricalData is relatively slow. To avoid doing it per
chunk, look up the values when the file is opened and just make a list
indexed by dump.

Also pack all the arguments that are going through from_block_function
into an object so that dask doesn't try to hash the individual arrays.