Split flag definitions out into katdal.flags
The FLAG_NAMES and FLAG_DESCRIPTIONS are then re-imported to h5datav2,
h5datav3 and visdatav4 (for compatibility), and also in the top-level
namespace. There are also definitions for the individual bits (e.g.
DATA_LOST_BIT = 3, DATA_LOST = 1 << DATA_LOST_BIT); for now I didn't
pull those into the top-level namespace because it's not obvious that
they represent flags when written as katdal.DATA_LOST, but
katdal.flags.DATA_LOST is self-explanatory.

In order for katdal.flags.DATA_LOST to work, I removed the code that
deletes the modules from the top-level namespace.

This addresses SR-1214 (although it does not introduce a `flags_raw`
array to datasets).