Load RDB files straight from archive
If given an HTTP URL, interpret it as the location of an RDB dump file
in an S3 object store. Use (or abuse) an S3ChunkStore object to access
it, which handles the necessary authorization. Download the RDB object
as a binary blob into memory and open it directly using the latest
katsdptelstate. Then also use this S3ChunkStore as the actual chunk
store in the absence of any overrides.

To enable the (ab)use of `S3ChunkStore`, upgrade the `_request` method
to its public API by dropping the underscore.

This addresses JIRA ticket SR-1443.