Sanitise HTTP requests to katstore
The katstore sensor info mechanism accepts Posix regular expressions
(via PostgreSQL) for sensor names, but we want to look up sensors by
their exact names. Avoid inadvertently returning multiple results
by stripping out all special characters in the sensor name and adding
regex anchors (^ and $).

The sensor name string is sanitised by insisting that it is a valid
Python identifier. This is precisely the escaped / normalised sensor
name used in KATCP resources and also in the katstore HTTP interface,
and coincidentally also in the sensor names found in telstate keys
(OK, maybe not so coincidental given the common creator).

An invalid Python identifier as sensor name currently triggers
a KeyError since it is akin to a missing sensor in SensorCache and
therefore might trigger further searches in future. On the other hand,
a ValueError feels more appropriate. The jury is still out on this.