Restore chunk store inference from RDB URL
If the RDB file is opened via archive URL, use that URL and
corresponding S3 credentials or token to access the chunk store. This
has been decreed since it is the expected behaviour 99% of the time.
On-site chunk stores will be accessed via whitelists only, thereby not
requiring credentials or tokens, which removes the concern that
prevented this shortcut. Overrides are still possible via URL parameters
(and separate RDB retrieval) for custom use cases.

This reverts the change in commit 98cbc8a and addresses JIRA ticket
SR-1443.