Clarify virtual pattern match and add tests
Clean up the regular expression processing of the virtual sensor
templates (this stuff dates back to 2011, so a bit crufty!).
Use string formatting to make it clearer what is going on.
Change `[^//]` to `[^/]`... Not sure why I had the slash twice (maybe
confusion with a backslash). Add unit tests to highlight the template
processing.