Speed up (u, v, w) but also change it slightly
The original calculation called `katpoint.Target.uvw` for each baseline,
triggered by a virtual sensor called `Corrprods/{antA}/{antB}/[uvw]`.
This would determine the (az, el) of the target as seen from antB.
This expensive calculation is then repeated for each baseline.

The new calculation uses the 'array' antenna to get the (az, el) of the
target once. This is cached as sensors that form a set of (u, v, w)
basis vectors (based on `katpoint.Target.uvw_basis`). A second layer of
virtual sensors caches the (u, v, w) coordinates between each antenna
and the 'array' antenna, as `Antennas/{ant}/[uvw]`. The katdal `u`, `v`
and `w` properties then loop over the baselines and take the difference
between the coordinates of the two antennas forming each baseline.
These are not cached but rather recalculated on the fly, potentially
saving a lot of memory (~180 MB for an 8-hour observation of the full
array).

Since all baselines use the same (az, el) from the 'array' antenna,
the answers differ slightly from the original case, where each baseline
had a slightly different (az, el). The coordinates differ by a few
millimetres, as reflected in the unit tests. This is potentially more
accurate, since the delay model obtains its w coordinates the same way.
The accuracy will be verified in JIRA ticket SR-1883.

The speed-up is on the order of the number of baselines, measuring
about 500x on a dataset with 60 antennas (1542913317).

One caveat is that this will break (u, v, w) on older MVF formats before
v4 that do not yet define an 'array' antenna. We'll need to remedy that
still.

This addresses JIRA ticket SR-1509 (and SR-1640).