Remove has_cal_product
This feature is sacrificed to allow cal product sensors to be virtual,
which will simplify the naming of L1 and L2 products a lot. While it is
possible to check if a virtual sensor template will match a sensor name,
this is no guarantee that the sensor will return valid data, as it may
in turn look for other sensors in the cache and not find them. The
situation is even messier with katstore sensors.

Instead, simply get a cal product when needed and deal with the fallout
if it's not there. Introduce a `skip_missing_products` parameter to
`calc_correction`, which skips cal products with missing sensors instead
of crashing. This is meant for the applycal='all' setting, where all
available products are applied.