Make cal stream name part of cal product
Up to now there has been only one calibration stream - 'cal' - and it's
hard-coded everywhere. Rather make the stream name part of the cal
product, which now has the format '<cal_stream>.<product_type>', like
'cal.G'. This will also disambiguate products of the same type provided
by different streams. If no stream is specified, we still assume 'cal'
for now to simplify unit tests etc.

Another benefit is that the cal `attrs` can lose the stream name prefix
to its keys, which will simplify future user-supplied products. This
was meant to be achieved via a telstate view on the 'cal' namespace.
Unfortunately this has the repercussion that the `attrs` in the data
source supplied to `VisibilityDataV4` has to be a telstate and not a
dict-like (since we need multiple views, one per cal stream). We'll
live with this for now and revisit it later.

This is the first slew of changes addressing JIRA ticket SR-1742.