Rename "L1" cal stream to "l1"
This bike shed was painted with the help of the pipeline team. Having
it lower-case makes it the same as "sdp_l0" and other related vis
stream names. It also distinguishes it from the Richter-case product
types appended to it to form full cal product names. Let's hope people
use sensible serif fonts to avoid confusion with "11"...