Prepare for multi-channel gain products
L2 introduces gains that are interpolated over time but also have
multiple channels ("IFs" in AIPS / Obit parlance). This requires a
rework of the code that maps the frequency axis of the calibration
solutions onto the frequency axis of the visibility data.

The original approach centred around the `calc_correction_per_corrprod`
function and only supported two versions: the `G` case that has a single
gain broadcast to all channels, and the `K` / `B` case that already has
a gain per visibility channel. We now need something intermediate. Each
vis channel will receive a gain from the cal channel that is closest in
frequency, but only during the `calc_correction` stage.

We need to store this channel map per cal stream, since it is produced
during the `add_applycal_sensors` phase. The relevant centre frequencies
(`cal_freqs` and `data_freqs`) and product type are all attributes that
are needed for this map, so transport them around instead. Finally add
the channel maps per cal product in `CorrectionParams` so that it can
be used to construct the dask correction array.

The design has become quite creaky (e.g. `add_applycal_sensors`
returning `cal_freqs` and the convoluted journey of `channel_maps`).
There is now a ticket (SR-1937) that aims to restore sanity to applycal
once the dust has settled.