Multi-channel gain products (GPHASE, GAMP_PHASE)
Introduce multi-channel gain products for the L2 stream. These perform
linear interpolation along the time axis, while picking the closest cal
channel for a given vis channel along the frequency axis. This behaviour
is triggered by the product type, regardless of the shape of the cal
product arrays - revisit that in SR-1937.

In the process the gain cal products are all promoted to 2-D arrays
(time by frequency), even if there is only a single gain for the entire
band.

This addresses JIRA ticket SR-1742.