Rework default cal stream
This was a bit of a mess, with a hard-coded 'cal' default in the
applycal module but a different default ('l1') at the visdatav4 level.

Remove all defaults from the applycal module, forcing it to use
fully qualified cal product names (like 'l1.G'). The _parse_cal_product
routine now raises a ValueError if this is not the case. Change the
parameter order of `get_cal_product` to a more natural one. Rework the
tests to use 'cal' explicitly. Finally, handle the actual default
stream setting at the highest level in visdatav4.