Generate a stable dask array name for corrections
The array name forms part of the dask key and will trigger unnecessary
recomputation if it differs for the same array. This is possible if
dask is executed in parallel on versions of Python with different
orderings of dictionary keys. Rather sort the cal products explicitly
and fix the order.