Allow gain interpolation to be target specific
Self-cal solutions are obtained on a specific source and should only be
interpolated to other data points of the same source. The gain
correction routine therefore needs to be aware of the target associated
with each dump (represented by the 'Observation/target_index' sensor).
For each gain channel, iterate over the targets and find all valid gains
on that target and interpolate them to other dumps of the same target.
Targets with no gain solutions will get invalid gains.

Only use this target-dependent interpolation with `GPHASE` and
`GAMP_PHASE` product types (associated with self-cal). Update the tests
so that `GPHASE` behaves like the eventual L2 stream (both multi-channel
and target-aware).

This addresses JIRA ticket SR-1742.