Return the final list of calibration products
This commit was inspired by a Jenkins unit test that failed because the
cal products were applied in a different order in the `applycal` module
(`calc_correction_per_corrprod`) and the unit test comparison
(`test_applycal.corrections_per_corrprod`) on Python 2.7, since they
involve the order of dictionary keys.

Instead of using an OrderedDict or constructing the correction dict in
the same way in both places, use the opportunity to return the actual
cal products in the order that they will be applied. This is useful info
because missing products may be skipped and the final list of products
to be corrected could be a small subset of the `cal_products` passed to
`calc_correction`. In addition, this info has been hidden up to now in
a `CorrectionParams` object wrapped inside a closure in a dask array.

Also remove dead code first introduced in 3cb072c - the `corrprods`
parameter of `corrections_per_corrprod`. This was meant for katdal
selection in the `DaskLazyIndexer` but that has all been reworked.