Create channel maps independent of product type
The calibration corrections (not to be confused with calibration streams
or products) come in three flavours:

- already interpolated onto the data frequencies (e.g. K and B),
- a single number for the entire band (e.g. G), or
- still stuck with the original cal frequency setup (e.g. GPHASE).

The number of frequency channels in each correction is therefore either
the number of data frequencies, one, or the number of cal frequencies,
respectively. Use this number to determine the appropriate channel map.
If data and cal have the same number of channels, check whether the
frequencies match up to decide between scenario 1 and 3 above.

Also add some minor documentation improvements requested in the PR
review.