Rework apply_weights_correction for unit test
The unit test for this function fails when NUMBA_DISABLE_JIT=1, with
the output disagreeing at the machine precision level. A slight tweak
to the calculation of the squared magnitude of the correction fixes
this and it is also faster in pure Python (10.5 vs 13 seconds on my
laptop). The Numba JIT performance remains the same (and that version
of the test also still passes).