Modify applycal tests to do sensor extraction
In order to fix the issue reported in SR-1575, first modify the tests
so that they could fail. Extract the data of the gain sensor and others
via :func:`sensor_to_categorical`, which allows specification of initial
values and other sensor properties. This mimics what will happen in
:class:`VisibilityDataV4`.

For simplicity, simply turn the gain event dump indices into timestamps
on integer seconds, with a dump period of 1.0. The main thing to check
will be the treatment of initial values. The K and B events are tweaked
so that the first event happens after dump 0. The default initial value
handler will move these events back to dump 0, which is currently the
expected behaviour. In contrast, moving the first G / GPHASE event to
a later dump crashes the tests (so we don't commit that part yet).