Set initial value of gain sensors to INVALID_GAIN
The gain sensors are unlikely to have a gain solution at dump 0, so the
categorical sensor extraction shifts the first proper gain solution back
to dump 0 to fill this void.

We want to avoid this for two reasons:
  - Linear interpolation of gains over time will be compromised.
  - The first gains will most likely land on a different target than
    intended, and this association can propagate it further in error.

One solution is to duplicate the first gain event at dump 0, but this
requires `allow_repeats=True` and a manual duplication process. Instead,
we set the initial value to `INVALID_GAIN`. When the gains are turned
into corrections, discard this initial placeholder and let the usual
extrapolation fill in the void instead. The placeholder is
incommensurate with the usual sensor values that are arrays, so we have
to strip it out up front. Also note that we have to check the identity
of the value since `INVALID_GAIN` contains NaNs (plus it's a singleton
anyway).

This addresses JIRA ticket SR-1575.