Add REST_FREQUENCY column to SOURCE sub-table
This makes CASA's exportuvfits happy, even though the SOURCE sub-table
is optional in an MS, and the REST_FREQUENCY column is again optional
in that table. It looks like exportuvfits only needs specific keywords
attached to the column, so add those based on @tony2heads's feedback.

This addresses JIRA tickets SR-1283 and SR-1811.