Improve populate_source_dict
Flesh out the docstring. Add comment lines to document each column.
Set NUM_LINES to 0, since we don't have any spectral lines for our
fields (setting it to the centre of the band was a lie). Remove data
for the non-existent SYSVEL column, but put an empty array in the
REST_FREQUENCY column instead, just in case someone attempts to read it
(the table crashes otherwise). We do not need the center_frequencies
parameter anymore, so remove from calling contexts, and fix
populate_ms_dict which no-one obviously uses anymore.