Add --applycal option to mvftoms.py
This passes a string of comma-separated calibration product names
through to `katdal.open`'s `applycal` parameter. The default is to
apply nothing. This also reports the products to be applied, although
this currently depends on a private `_applycal_products` attribute.
We'll fix that in a separate PR.

This addresses JIRA ticket SR-1948.