Merge open_main into open_table and simplify
The current `open_table()` suffers from warts linked to the now defunct
casapy support. Since an initialiser will either return the object or
raise an exception (RuntimeError in this case), there is no chance of
a None popping out. Rather remove all those checks and let exceptions
happen if they happen.

While we are at it, merge `open_main` into `open_table` since there is
now no difference between the two. I dabbled with the idea of dropping
`open_table` altogether in favour of `tables.table`, but this way the
dependency on casacore is contained in `ms_extra`. This would be useful
if we ever contemplate restoring casapy support.

Use the convention that the table and sub-table names are separated by
'::' instead of forcing a filesystem relationship.