Add applycal_products attribute to DataSet
This exposes the final set of calibration products that will be applied
to the data. It is currently only supported in VisibilityDataV4, where
it returns the cal products found and selected by the `applycal`
parameter. It should currently only be set in the initialiser (i.e. it
is a read-only property), but I've left it as is as things may change
when user-supplied cal products are added.