Streamline S3ChunkStore construction
Get rid of the superfluous layers of `from_url` and `_from_url` which
caused a lot of code duplication and confusion (as lamented in #235).
All construction used to go via `S3ChunkStore.from_url` and no-one
actually called `S3ChunkStore.__init__` directly... Therefore move
`from_url` over to the initialiser. Everything else remains the same
(for now).

Do the smoke test inside the initialiser instead. This store ping is
the only bit that needs an enforced timeout via queue and thread since
it is the first attempted network request.