Remove hard timeout machinery aimed at slow DNS
The initial idea was to enforce a hard timeout on the first attempt to
connect to the S3 chunk store server. This was mostly to avoid
interminable hangs while looking up the server name via DNS in some
pathological setups. Since socket.gethostbyname does not play nicely,
we needed the complication of a second thread and a queue to relay back
any errors. The thread could potentially hang around for the rest of
the script lifetime.

Remove all of this and don't try to work around broken DNS setups.
It's just not worth it and can be misleading, as witnessed in JIRA
ticket SR-1961.