Reinterpret HTTP statuses in S3ChunkStore
A 401 (Unauthorized) now results in a more specific AuthorisationFailed
exception.

The current MeerKAT Ceph RADOS Gateway actually returns a 403 for
missing chunks and other objects (like the 'complete' marker). This is
due to a bug in the interaction with bucket policy - see the discussion
at https://tracker.ceph.com/issues/38638 and the resulting PR at
ceph/ceph#27309 for a possible future resolution. Turn both 403 and 404
into ChunkNotFound if the object in question was expected to live within
a bucket. Missing buckets are actually worse and leads straight to
StoreUnavailable.

This addresses JIRA ticket SPR1-181.