Improve timeouts and retries
Allow custom connect and read timeouts. For starters, let the read
timeout be 300 seconds as before, to cater for latency in getting the
data out of Ceph and through the proxy. Change the connect timeout to
30 seconds, however, to allow for faster failure if the server is not
up. This will play the role of the "smoke test" in future.

Add a `retries` parameter to `S3ChunkStore` which can be a full-blown
Retry object for a completely custom retry strategy. The new default
retry strategy also retries 50x server errors within a 5-minute backoff
period, since these are associated with a strained Ceph gateway and
should be temporary. If there are persistent 50x errors while getting
an object from the store and the retries fail, treat it the same way as
a 404 "Not Found" error, which are typically mapped to `ChunkNotFound`.

Let the unit tests have quicker timeouts and fewer retries (or none at
all, in some cases).

This addresses some of the main issues of Jira ticket SPR1-181.