Test server errors with proxy handler suggestions
Inspired by httpbin.org, slipstream a suggestion to the token-based
HTTP proxy handler inside the requested URL path to simulate various
error conditions on the server side. For example, if you GET
http://127.0.0.1:12345/bucket/please-respond-with-502/x/00000.npy
this will instruct the proxy handler to respond with a status code of
502 after a nominal 0.1 seconds instead of contacting the minio server.

The error condition can also go away after a fixed period of time, by
changing the suggestion to .../please-respond-with-502-for-2-seconds/...
This strips off the suggestion from the true path and performs the
intended GET on minio once a retry appears more than 2 seconds after the
initial request.

Finally, check that a status-based retry strategy does the intended
thing.