Replace home-brewed JWT routines with PyJWT
In the interest of maintainability, use a well-trodden package for
decoding JSON web tokens instead of reinventing the wheel. NIHSH.

Only do the minimum checks that will detect truncated or extended
tokens, tokens with corrupted characters and expired tokens, and tailor
the error messages to suggest this to the user (since this is at the
coal face). Similarly, `encode_jwt` only exists to support the unit
tests. An additional benefit is that the `InvalidToken` exception is
now a subclass of `AuthorisationFailed`, instead of chaining them
together artificially.