Remove S3 server smoke test
The strict token prefix check added in c0b842c now prevents S3 root
directory access, even though the proxy server allows anyone to list
the root directory as a special case (but this is not reflected in the
allowed prefix claim in the token). This breaks the smoke test, but
that's OK since the smoke test has outlived its usefulness anyway.
This special proxy rule is also troublesome since it always succeeds,
even with a bad or missing token.

While the smoke test served as some kind of "knock on the door," to
use the parlance of SPR1-181, we still have to connect to the bucket
to get at the objects in a second request, which might then still fail.
It also prevents katdal from opening a local RDB file to inspect
metadata without S3 access, requiring the cumbersome `chunk_store=None`
parameter to work.

A better solution is to make connection timeouts quicker in general -
the thrust of the next PR in this series. Then each request effectively
"knocks on the door" first.