Do away with _raise_for_status
In preparation for retrying reset connections, merge the standalone
`_raise_for_status` function into the `S3ChunkStore.request` method,
since it will need access to more attributes of the store. This has
the side effect of centralising all HTTP requests in this method
(the `create_array` bucket PUT was the last holdout).

There is also no need to raise an HTTPError exception, only to chain it
to a standard ChunkStore one directly after. The only downside is an
assumption that the response reason is a string and not bytes
(something that `requests.Response.raise_for_status` took care of).
Our Ceph gateway seems to honour this assumption.

Improve the error reporting inside `S3ChunkStore.request` to be more
like `ChunkStore._standard_errors`. The latter is still retained for
exceptions emanating from within the with-block that will process a
successful HTTP response. Allow some HTTP statuses to be tolerated
via the `ignored_errors` parameter, which is useful for the bucket PUT
(the reason it was a holdout before).