Split get_chunk into two parts
In preparation for retrying reset connections (which will require
multiple HTTP requests in one `get_chunk` call), split out the
lower-level request part to a new method called `_get_chunk`. A side
benefit of this new method is that it simplifies interactive use and
testing of the S3ChunkStore: `store._get_chunk(url)` is all you need
to do to get a chunk (no need to fabricate slices, dtypes, etc).