Detect truncated data and raise TruncatedRead
Introduce new `TruncatedRead` exception that represents truncated
content in a response from S3 server. Monkey-patch NumPy so that this
truncation can also be detected in the magic and header parts of the
content. This is needed because all issues with reading an array are
lumped together as ValueErrors (and that's why `TruncatedRead` is a
`ValueError` too). Extend the existing unit test for short reads to
check that the appropriate exception is raised in any part of the
content.