Do basic retry of truncated reads
If the S3 chunk store responds with fewer bytes than expected to a GET
request, this is typically because the TCP connection was reset due to
server overload. We would like to retry this and eventually fail with
ChunkNotFound instead of crashing.

The current retry strategy is very simplistic and hardcoded just to get
going. It will be linked to the 50x strategy and combined with it in a
future commit.

Extend the HTTP proxy in the unit tests to test truncated reads.

This addresses JIRA tickets SR-1529 and SPR1-173, as well as umbrella
ticket SPR1-181.