Improve debug logs in S3ChunkStore tests
Log the time since the first attempt at a given request, which is useful
for debugging the fault injection timeline of the proxy handler.
Simplify the lookup of these initial times by using the entire request
line as a key, and wield the wondrous dict.setdefault as suggested by
Bruce.