Move cache of request timestamps to server object
Instead of using a global module-level attribute, move the dict of
initial request times to the HTTP server object. This object is
recreated with every test case, which conveniently clears this cache
too, unlike the module-level attribute which potentially leaked state
across tests. The server object is accessible within the handler object.
The HTTP handler object itself is not suitable for storing the time
cache because it gets recreated with *every* request.