Add pause-in-the-middle test
This test lets the proxy server write half a chunk, pause, and then
write the rest. The pause is nominally 0.1 seconds (I also tried 0.3
seconds), which is shorter than the read timeout of 0.4 seconds but long
enough to maybe trick the client into thinking the read is done for now.

This tests whether the client read implementation is blocking (only
return all requested bytes or EOF) or semi-blocking (return a partial
read). If the latter, it could trigger false read truncation errors.

The pausing is done via a new proxy suggestion ('please-pause-read'),
instead of doing it on every read. I tried the latter but ran into
extra connection errors and broken pipes on the server side (and it's
slower).

The test passes, indicating blocking behaviour.