Make mark_complete work on whole bucket again
PR #274 broke `mark_complete` when run on a whole bucket, because it
calls `create_array` which now crashes if the "array_name" doesn't have
a separator.

This keeps the behaviour for `create_array`, but adds a backdoor
`_create_bucket` so that `mark_complete` can once again put a completion
marker at the bucket level.