Increase S3 connection timeouts in unit tests
Since we are connecting to a local HTTP proxy server running on a
second Python thread in the same process, things can get a bit jerky.
Increase the connect timeout substantially to 5 seconds. Also redo the
unresponsive server test and give it its own (short) timeout without
retries to speed it up (no need to wait for 10 seconds).

This addresses JIRA ticket SPR1-364.