Improve error message for expired tokens
The current error message is the default one provided by PyJWT which
reads "Signature has expired" - but what is a "signature"? To help the
end user, call it a token instead and include the expiry date/time for
handy reference. Also mention that the correct course of action is to
obtain a new one (thanks Ruby!).

This addresses JIRA ticket SR-1971.