Add flux calibration to applycal
This applies a flux scale to those gains of the "G" calibration product
that are associated with gain calibrator targets with a valid flux
density in the `gaincal_fluxes` dict. More specifically, the gains are
multiplied by the square root of the flux, which turns the default
flux model value of 1.0 into the appropriate flux for the gain
calibrator. The fluxes are obtained from the `measured_flux` telstate
key produced by the cal pipeline but it can also be overridden via
`katdal.open` or disabled by specifying an empty dict.

This is equivalent to the final step of the AIPS GETJY and CASA
fluxscale tasks.

This addresses JIRA tickets SPR1-67 and SPAZA-27.