Look for cal attributes in all namespaces
The `measured_flux` telstate key added by SPR1-67 lives in the CB_cal
namespace and not in the cal namespace like the other attributes. In
this regard it is more like a cal product than an attribute, although
that classification has way more downsides.

Extend the namespaces to search for L1 cal attributes by creating a
"relative" telstate view. This appends "cal" to all existing namespaces
in the main telstate (which includes CB as a namespace). The reasoning
behind this is that the telstate view should be preconfigured by the
data source and not the `VisibilityDataV4` class, which merely
interprets that view. The latter should not have to know the CBID or
the exact order of prefixes in the main telstate. By adding "cal" to
all namespaces, we achieve this separation of concerns.

Coincidentally, the L2 cal attribute set is already a relative telstate
view. The only difference is that it included the top-level '' prefix,
but this can be safely removed to use the same _relative_view function
as L1.

The relative view functionality should eventually move into
katsdptelstate proper, as a relative=True parameter to
`TelescopeState.view`.