Avoid using has_array in ChunkStoreVisFlagsWeights
Instead, just try to load things, and if any of them fail, apply flags.
This has a few advantages:

1. chunkstore_s3 is non-deterministic, because it will treat a 403 or
50x error as a missing chunk even if the bucket listing included the
chunk. This meant that even when loading vis, flags, weights jointly,
the vis/weights could be missing and the flags wouldn't reflect this (or
if the flags got a 50x, it would return zero flags). With this change,
joint loading should always return consistent results.

2. Getting the list of all chunks could add a lot of time to
`katdal.open`, which now disappears.

A synthetic benchmark (with no applycal) shows that loading drops from
about 3.3GB/s to about 3GB/s, but that's still faster than the 25Gb/s
network can deliver the data.

This is part 1, which just removes the calls to has_array. Part 2 will
remove all the has_array interfaces and the dependence on defusedxml.