Speed up unique_in_order for large ndarrays
Instead of comparing each array to a previous array, hash each array and
look up the hash in a dictionary. While cityhash is not strictly
required, it is much, much faster than the default (sha1) if no fast
3rd-party hasher is found, so it's added to the requirements so that
users won't need to know to install it separately for good performance.

This takes at least 30 seconds off the time to open a large file when
using applycal.