Refactor caching for computed cal products
Instead of Calibration/Products/cal/B<n> which have to be stitched
together across frequencies to form the bandpass solution, there is now
a virtual Calibration/Products/cal/B which handles the stitching
internally. This allows it to be cached once extracted.

The unit tests still need to be updated to match.