Fix spurious warnings due to NaN != NaN
remove_duplicates_and_invalid_values would compare values even when all
timestamps are unique. Apart from being inefficient, it also led to
spurious warnings when the arrays contained NaNs and hence did not
compare equal to themselves.

There is still an issue that if there *are* duplicate timestamps but the
values are the same (and contain NaNs) then the warning will be issued
unnecessarily. But that is a much less common case.