Use MutableMapping for SensorCache
This provides more predictable/controllable behaviour compared to
inheriting directly from dict. I also made some fixes to
ConcatenatedSensorCache regarding iteration and length (previously it
implemented iterkeys, which didn't work at all in Python 3 and also made
the assumption that the underlying caches had exactly the same keys).

Querying a sensor cache can cause new keys to appear in the underlying
dict (as virtuals are expanded), so it was necessary to make some
choices). I decided that __contains__ will only consider instantiated
keys, not virtuals, making it consistent with `keys()` and `len` (and
the previous implementation of SensorCache), but inconsistent with
`__getitem__`.

One user-visible change in SensorCache is that `items()` and `.values()`
will now return extracted data, because MutableMapping implements them
in terms of `__getitem__`.

Closes SPR1-445.