Fix filtering of sensor events with invalid status
After sorting the sensor timestamps, the values are updated accordingly
but not the statuses, which is a bug. Normally this is not an issue
because timestamps tend to be chronological anyway. The somewhat
insidious interaction between value and sample/update timestamps may
however occasionally cause a later event to be sampled before an earlier
event. An example is the cbf_1_wide_input_labelling sensor during CBID
1588667937 which triggered this fix. Add a test to check this corner
case and also exercise the various code paths in the filter function.